+++ lib/mix/tasks/compile/grid_codec.ex
@@ -231,7 +231,8 @@ defmodule Mix.Tasks.Compile.GridCodec do
     encode_clauses =
       Enum.map(codecs, fn %{module: mod} ->
         quote do
-          def encode(%unquote(mod){} = struct, opts), do: unquote(mod).encode(struct, opts)
+          def encode(%{__struct__: unquote(mod)} = struct, opts),
+            do: unquote(mod).encode(struct, opts)
         end
       end)